requirements.txt
Flask>=3.1.1
flask-cors>=6.0.0
flask-limiter>=3.5.0
requests>=2.31.0
beautifulsoup4>=4.13.0
lxml>=5.0.0
urllib3>=2.0.0
python-dotenv>=1.1.0
gunicorn>=23.0.0
supabase>=2.15.1
